python-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gotrue==2.12.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
instructor==1.7.9
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-core==0.3.51
langchain-google-genai==2.0.10
langchain-mistralai==0.2.10
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.2.76
langgraph-checkpoint==2.0.24
langgraph-sdk==0.1.61
langsmith==0.3.30
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
mistralai==1.6.0
multidict==6.4.3
ngrok==1.4.0
numpy==2.2.4
openai==1.72.0
openpyxl==3.1.5
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
portia-sdk-python==0.1.14
postgrest==1.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
websockets==14.2
win32_setctime==1.2.0
yarl==1.19.0
zstandard==0.23.0
